final commit project0 demo

+++ src/main/java/com/training/pms/ProductApp.java
@@ -1,7 +1,11 @@
 package com.training.pms;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Scanner;
+
 import com.training.pms.dao.ProductDAO;
 import com.training.pms.dao.ProductDAOImpl;
 import com.training.pms.model.Product;
@@ -9,9 +13,12 @@ import com.training.pms.model.Product;
 public class ProductApp {
 	Scanner scanner = new Scanner(System.in);
 	int choice = 0;
-	//Animal a = new Cat();
+	// Animal a = new Cat();
 	ProductDAO productDAO = new ProductDAOImpl();
 	Product product = new Product();
+	boolean result;
+	List<Product> products = new ArrayList<Product>();
+
 	public void startProductApp() throws IOException {
 
 		// declaring local variables for input
@@ -19,7 +26,7 @@ public class ProductApp {
 		String productName = null;
 		int quantityOnHand = 0;
 		int price = 0;
-		
+
 		while (true) {
 			System.out.println("=================================================================");
 			System.out.println("P R O D U C T    -     APP      M E N U");
@@ -35,37 +42,48 @@ public class ProductApp {
 			System.out.println("Please enter your choice : ");
 			choice = scanner.nextInt();
 			System.out.println("=================================================================");
-			
+
 			switch (choice) {
 			case 1:
 				// add product section
 				System.out.println("WELCOME TO ADD PRODUCT SECTION ");
-				
-
 				// take input from user to add a product
 				System.out.println("Please enter product id :");
 				productId = scanner.nextInt();
 
+				if (productDAO.isProductExists(productId)) {
+					System.out.println("Product with product id : " + productId
+							+ " already exists, please try with another product id");
+					continue;
+				}
 				System.out.println("Please enter product name :");
 				productName = scanner.next();
-
 				System.out.println("Please enter product qoh :");
 				quantityOnHand = scanner.nextInt();
-
 				System.out.println("Please enter product price :");
 				price = scanner.nextInt();
 
 				product = new Product(productId, productName, quantityOnHand, price);
 				// call dao layer to save the product
-				productDAO.addProduct(product);
-				System.out.println("Congratulations, your product : " + productName + " saved successfully");
-
+				result = productDAO.addProduct(product);
+				if (result)
+					System.out.println("Congratulations, your product : " + productName + " saved successfully");
+				else
+					System.out.println("Sorry your product cannot be saved");
 				break;
-				
+
 			case 2:
 				System.out.println("Please enter product id to delete :");
 				productId = scanner.nextInt();
-				productDAO.deleteProduct(productId);
+
+				if (productDAO.isProductExists(productId)) {
+					productDAO.deleteProduct(productId);
+					System.out.println("Product with product id : " + productId + " deleted successfully");
+				} else {
+					System.out.println(
+							"Product with product id : " + productId + " does not exists, hence cannot be deleted");
+
+				}
 				break;
 			case 3:
 				// update product section
@@ -74,6 +92,12 @@ public class ProductApp {
 				// take input from user to update a product
 				System.out.println("Please enter product id to update :");
 				productId = scanner.nextInt();
+
+				if (!productDAO.isProductExists(productId)) {
+					System.out.println(
+							"Product with product id : " + productId + " does not exists, so cannot be updated");
+					continue;
+				}
 
 				System.out.println("Please enter new product name :");
 				productName = scanner.next();
@@ -84,35 +108,71 @@ public class ProductApp {
 				System.out.println("Please enter new product price :");
 				price = scanner.nextInt();
 
-				 product = new Product(productId, productName, quantityOnHand, price);
+				product = new Product(productId, productName, quantityOnHand, price);
 				// call dao layer to save the product
 				productDAO.updateProduct(product);
 				System.out.println("Congratulations, your product : " + productName + " updated successfully");
 
 				break;
-				
+			case 4:
+				System.out.println("Please enter product id to search :");
+				productId = scanner.nextInt();
+
+				if (productDAO.isProductExists(productId)) {
+					Product temp = productDAO.searchByProductId(productId);
+					System.out.println(temp);
+				} else {
+					System.out.println(
+							"Product with product id : " + productId + " does not exists, hence cannot be deleted");
+
+				}
 			case 5:
 				System.out.println("Please enter product name to search :");
 				productName = scanner.next();
-				productDAO.searchByProductName(productName);
+				products = productDAO.searchByProductName(productName);
+				if (products.size() == 0) {
+					System.out.println("No products matching your criteria");
+					continue;
+				}
+				printProductDetails(products);
 				break;
 			case 6:
-					productDAO.printAllProducts();
-					break;
+				if (products.size() == 0) {
+					System.out.println("No products");
+					continue;
+				}
+				products = productDAO.getProducts();
+				printProductDetails(products);
+
+				break;
 			case 7:
 				System.out.println("Please enter product price (lower) :");
 				int lowerprice = scanner.nextInt();
 				System.out.println("Please enter product price (upper) :");
 				int upperprice = scanner.nextInt();
-				productDAO.searchProductByPrice(lowerprice, upperprice);
-				break;	
-					
-			case 9: System.out.println("Thanks for using my product app");
-					System.exit(0);
+				products = productDAO.searchProductByPrice(lowerprice, upperprice);
+				if (products.size() == 0) {
+					System.out.println("No products matching your criteria");
+					continue;
+				}
+				printProductDetails(products);
+				break;
+
+			case 9:
+				System.out.println("Thanks for using my product app");
+				System.exit(0);
 			default:
 				System.out.println("Invalid Choice , Please enter (1-6) or 9 to EXIT");
 				break;
 			}
+		}
+	}
+
+	public void printProductDetails(List<Product> products) {
+		Iterator<Product> iterator = products.iterator();
+		while (iterator.hasNext()) {
+			Product temp = iterator.next();
+			System.out.println(temp);
 		}
 	}
 }
